removed param jsons. sorted functions
