--- parameter-file-application.json
@@ -1,10 +1,0 @@
-{
-    "common": {
-        "fsl_bin_path": "/panfs/roc/msisoft/fsl/6.0.2/bin/"
-    },
-
-    "bibsnet": {
-        "model": "3d_fullres",
-        "nnUNet_predict_path": "/home/support/public/torch_cudnn8.2/bin/nnUNet_predict"
-    }
-}
