--- parameter-file-container.json
@@ -1,10 +1,0 @@
-{
-    "common": {
-        "fsl_bin_path": "/opt/fsl-6.0.5.1/bin/"
-    },
-
-    "bibsnet": {
-        "model": "3d_fullres",
-        "nnUNet_predict_path": "/opt/conda/bin/nnUNet_predict"
-    }
-}
