--- src/param-types.json
@@ -1,11 +1,0 @@
-{
-    "common": {
-        "fsl_bin_path": "existing_directory_path"
-    },
-
-    "bibsnet": {
-        "model": "str",
-        "nnUNet_predict_path": "existing_file_path" 
-    }
-
-}
